// src/api/stetsom/model/partnerLocationType.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type PartnerLocationType =
  (typeof PartnerLocationType)[keyof typeof PartnerLocationType];

export const PartnerLocationType = {
  REPRESENTATIVE: "REPRESENTATIVE",
  SERVICE_CENTER: "SERVICE_CENTER",
} as const;