// src/api/stetsom/model/postApiPartnerLocationsImportWorkbook200.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiPartnerLocationsImportWorkbook200Report } from "./postApiPartnerLocationsImportWorkbook200Report";
import type { PostApiPartnerLocationsImportWorkbook200SheetsItem } from "./postApiPartnerLocationsImportWorkbook200SheetsItem";

export type PostApiPartnerLocationsImportWorkbook200 = {
  sheets: PostApiPartnerLocationsImportWorkbook200SheetsItem[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  totalRows: number;
  warnings: string[];
  errors: string[];
  report: PostApiPartnerLocationsImportWorkbook200Report;
};